requirements.txt
# Core Dependencies
requests>=2.32.0
PyMySQL>=1.1.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.30
pandas>=2.2.0

# YouTube API
google-api-python-client>=2.141.0

# Sentiment Analysis & NLP
vaderSentiment>=3.3.2
textblob>=0.17.1
nltk>=3.8.1

# Bot Detection & Text Processing
scikit-learn>=1.4.0
numpy>=1.24.0
regex>=2023.12.25

# Data Visualization
plotly>=5.24.0
altair>=5.3.0
vega_datasets>=0.9.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Jupyter & Notebooks
nbformat>=5.10.4
nbconvert>=7.16.4
jupyter-client>=8.6.3
ipywidgets>=8.1.0
jupyter>=1.0.0
jupyterlab>=4.0.0

# Testing & Quality
pytest>=8.2.0
pytest-timeout>=2.3.1
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0

# Development Tools
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0

# Security & Monitoring
bandit>=1.7.5
safety>=3.0.0

# CLI & Utilities
typer>=0.12.5
rich>=13.7.0
tqdm>=4.66.0

# Date & Time
python-dateutil>=2.8.2

# HTTP & API
urllib3>=2.0.0
certifi>=2023.11.17
